feat(bundler): Manage saved asset files
bundler.define returns the destination paths so pack can manage the output files.
