+++ docs/js/docs-setup.js
@@ -48,7 +48,7 @@ NG_DOCS={
       "type": "overview",
       "moduleName": "Client-Side Bundler",
       "shortDescription": "Client-Side Bundler",
-      "keywords": "action additional arguments asset assets aware based browserify bulk bundle bundler bundlers bundling callback called cb changes client client-side code complete completely config contents create css current custom default define definition described development directory discuss dropped early ensureassetfolder experimental features file files function future html implement implementation implemented implementing individual jade js jspm lacks load method methods minification move named newer next_arg object objective optimisations options opts overview pack pass passed production referenced responsibilities return saved scss separate served server ss starting step supported templates text tmpl tutorials types var view views webpack webpackbundler worker"
+      "keywords": "action additional arguments asset assets aware based browserify bulk bundle bundler bundlers bundling callback called cb changes client client-side code complete completely config contents create css current custom default define definition described destsfor development directory discuss dropped early experimental features file files function future html implement implementation implemented implementing individual jade js jspm lacks load method methods minification move named newer next_arg object objective optimisations options opts overview pack pass passed production referenced responsibilities return saved scss separate served server ss starting step supported templates text tmpl tutorials types var view views webpack webpackbundler worker"
     },
     {
       "section": "tutorials",
