+++ lib/client/bundler/default.js
@@ -18,22 +18,6 @@ function includeFlags(overrides) {
   return includes;
 }
 
-function deleteOldFiles(clientDir) {
-  var filesDeleted, numFilesDeleted;
-  numFilesDeleted = 0;
-  filesDeleted = fs.readdirSync(clientDir).map(function(fileName) {
-    return fs.unlinkSync(path.join(clientDir, fileName));
-  });
-  return filesDeleted.length > 1 && log('✓'.green, '' + filesDeleted.length + ' previous packaged files deleted');
-}
-
-function mkdir(dir) {
-  if (!fs.existsSync(dir)) {
-    return fs.mkdirSync(dir);
-  }
-}
-
-
 module.exports = function(ss,client,options){
   var bundler = {
     define: define,
@@ -47,8 +31,7 @@ module.exports = function(ss,client,options){
     pack: {
       js: packJS,
       css: packCSS
-    },
-    ensureAssetFolder: ensureAssetFolder
+    }
   };
 
   return bundler;
@@ -79,11 +62,11 @@ module.exports = function(ss,client,options){
     client.paths = paths;
     client.includes = includeFlags(paths.includes);
 
-    return client;
+    return ss.bundler.destsFor(ss,client,options);
 	}
 
   function load() {
-    this.description = ss.bundler.descriptionFor(ss,client,options);
+
   }
 
   function assetJS(path, opts, cb) {
@@ -118,17 +101,6 @@ module.exports = function(ss,client,options){
 
   function packCSS(postProcess) {
     ss.bundler.packAssetSet('css', options.dirs.css, client, bundler, postProcess);
-  }
-
-  //TODO move to index.js
-  function ensureAssetFolder() {
-
-    // Prepare folder
-    mkdir(this.description.containerDir);
-    mkdir(this.description.dir);
-    if (!(options.packedAssets && options.packedAssets.keepOldFiles)) {
-      deleteOldFiles(this.description.dir);
-    }
   }
 };
