+++ lib/client/bundler/index.js
@@ -8,6 +8,7 @@ var fs = require('fs'),
     system = require('../system'),
     view = require('../view'),
     wrap = require('../wrap'),
+    magicPath = require('../magic_path'),
     uglifyjs = require('uglify-js'),
     jsp = uglifyjs.parser,
     pro = uglifyjs.uglify;
@@ -30,10 +31,10 @@ exports.define = function defineBundler(ss,client,args,options) {
 
   if (typeof pathsOrFunc === "function") {
     bundlers[name] = pathsOrFunc(ss,options);
-    bundlers[name].define(client, args[2], args[3], args[4], args[5]);
+    bundlers[name].dests = bundlers[name].define(client, args[2], args[3], args[4], args[5]);
   } else {
     bundlers[name] = require('./default')(ss,client,options);
-    bundlers[name].define(args[1]);
+    bundlers[name].dests = bundlers[name].define(args[1]);
   }
 };
 
@@ -73,7 +74,12 @@ exports.pack = function pack(ss, client, options) {
 
   log(('Pre-packing and minifying the \'' + client.name + '\' client...').yellow);
   
-  bundler.ensureAssetFolder();
+  // Prepare folder
+  mkdir(bundler.dests.containerDir);
+  mkdir(bundler.dests.dir);
+  if (!(options.packedAssets && options.packedAssets.keepOldFiles)) {
+    deleteOldFiles(bundler.dests.dir);
+  }
 
   // Output CSS  
   bundler.pack.css(function(files) {
@@ -100,10 +106,25 @@ exports.pack = function pack(ss, client, options) {
 
   // Output HTML view  
   return view(ss, client, options, function(html) {
-    fs.writeFileSync(bundler.description.paths.html, html);
-    return log.info('✓'.green, 'Created and cached HTML file ' + bundler.description.relPaths.html);
+    fs.writeFileSync(bundler.dests.paths.html, html);
+    return log.info('✓'.green, 'Created and cached HTML file ' + bundler.dests.relPaths.html);
   });
 };
+
+
+function deleteOldFiles(clientDir) {
+  var numFilesDeleted = 0,
+      filesDeleted = fs.readdirSync(clientDir).map(function(fileName) {
+        return fs.unlinkSync(path.join(clientDir, fileName));
+      });
+  return filesDeleted.length > 1 && log('✓'.green, '' + filesDeleted.length + ' previous packaged files deleted');
+}
+
+function mkdir(dir) {
+  if (!fs.existsSync(dir)) {
+    return fs.mkdirSync(dir);
+  }
+}
 
 // API for implementing bundlers
 
@@ -134,7 +155,7 @@ exports.minifyJSFile = function minifyJSFile(originalCode, fileName) {
   return minifiedCode;
 };
 
-exports.descriptionFor = function(ss,client,options) {
+exports.destsFor = function(ss,client,options) {
   var containerDir = path.join(ss.root, options.dirs.assets);
   var clientDir = path.join(containerDir, client.name);
 
@@ -162,9 +183,9 @@ exports.packAssetSet = function packAssetSet(assetType, dir, client, bundler, po
     paths = client.paths[assetType];
 
   function writeFile(fileContents) {
-    var fileName = bundler.description.paths[assetType];
+    var fileName = bundler.dests.paths[assetType];
     fs.writeFileSync(fileName, postProcess(fileContents));
-    return log.info('✓'.green, 'Packed', filePaths.length, 'files into', bundler.description.relPaths[assetType]);
+    return log.info('✓'.green, 'Packed', filePaths.length, 'files into', bundler.dests.relPaths[assetType]);
   }
 
   function processFiles(fileContents, i) {
@@ -214,7 +235,6 @@ exports.formatKb = function formatKb(size) {
 // is wrapped in a module wrapper (to keep vars local and allow you to require() one file in another).
 // The 'system' directory is a special case - any module placed in this dir will not have a leading slash
 exports.wrapCode = function wrapCode(code, path, pathPrefix, options) {
-  var modPath, sp, i;
   var pathAry = path.split('/');
 
   // Don't touch the code if it's in a 'libs' directory
@@ -223,9 +243,10 @@ exports.wrapCode = function wrapCode(code, path, pathPrefix, options) {
   }
 
   if (pathAry.indexOf('entry.js') === -1 && options && options.browserifyExcludePaths) {
-    for(i in options.browserifyExcludePaths) {
-      if (options.browserifyExcludePaths.hasOwnProperty(i)) {
-        if ( path.split( options.browserifyExcludePaths[i] )[0] === '' ) {
+    //TODO is this an array? should be revised
+    for(var p in options.browserifyExcludePaths) {
+      if (options.browserifyExcludePaths.hasOwnProperty(p)) {
+        if ( path.split( options.browserifyExcludePaths[p] )[0] === '' ) {
           return code;
         }
       }
@@ -234,19 +255,18 @@ exports.wrapCode = function wrapCode(code, path, pathPrefix, options) {
 
   // Don't add a leading slash if this is a 'system' module
   if (pathAry.indexOf('system') >= 0) {
-    modPath = pathAry[pathAry.length - 1];
-    return wrap.module(modPath, code);
+    return wrap.module(pathAry[pathAry.length - 1], code);
   } else {
 
     // Otherwise treat as a regular module
-    modPath = options.globalModules? pathAry.join("/") : pathAry.slice(1).join('/');
+    var modPath = options.globalModules? pathAry.join("/") : pathAry.slice(1).join('/');
 
     // Work out namespace for module
     if (pathPrefix) {
 
       // Ignore any filenames in the path
       if (pathPrefix.indexOf('.') > 0) {
-        sp = pathPrefix.split('/');
+        var sp = pathPrefix.split('/');
         sp.pop();
         pathPrefix = sp.join('/');
       }
