--- lib/client/bundler/pack.js
@@ -1,14 +1,0 @@
-'use strict';
-
-var fs = require('fs'),
-    path = require('path'),
-    log = require('../../utils/log'),
-    magicPath = require('../magic_path'),
-    view = require('../view');
-
-module.exports = function(ss, client, bundler, options) {
-
-
-  return ;
-};
-
